chore(campaigns): delete the unreachable classic DesignEditor (P2-5)
Proof chain, re-verified on current main: the workspace mounts DesignEditor
only when studioSupportsCampaign() is false; that flag is exactly
`type !== 'guided_review'`; and DesignEditor's first branch dispatches
guided_review to GuidedReviewDesigner — so ClassicDesignEditor, the
StudioV2Notice branch, and every panel under src/components/campaigns/editor/
could never render. The only other mount (dev-only GuidedReviewDemo) is also
guided_review. The one thing keeping it compiled was the Studio Quiz panel
leeching STARTER_QUIZ/BLANK_QUIZ from the dead editor/QuizPanel.

- Hoist STARTER_QUIZ + BLANK_QUIZ verbatim into src/lib/quizTemplates.js
  (beside quizScoring/quizDisplay); repoint the Studio panel + 2 studio tests
- Reduce DesignEditor.jsx (304 lines) to the guided_review dispatcher;
  non-guided mounts render null (backend 409s classic writes over v2 anyway)
- Delete src/components/campaigns/editor/ — ContentPanel, DesignPanel,
  FieldOrderEditor, LayoutPanel, MarketplacePanel, PreviewFrame, QuizPanel,
  constants.js (StudioV2Notice lived inside DesignEditor.jsx, not editor/ —
  deleted with its branch; zero external importers)
- Delete DesignEditorV2Guard.test.jsx (covered only deleted branches);
  rewrite DesignEditor.test.jsx as dispatcher tests (guided mounts + props
  forwarded, non-guided/v2 render nothing)
- previewFixtureInertness.test.jsx checked: pure live-funnel coverage
  (CampaignSignupForm/CampaignQuiz/CampaignPageRenderer) — untouched

Kept: GuidedReviewDesigner, guidedReviewDefaults, GuidedReviewPage (live for
customers via LeadCapture + public Preview); AdminCampaignWorkspace.jsx and
studioFlag.js are byte-untouched, so the Studio path for normal campaigns
cannot have changed.

Verified: vitest 2119 green (delta = deleted classic tests, net of 3 new
dispatcher tests; repointed studio quiz suites pass); mktr+redeem+ops builds
green; Playwright on the dev server renders the full GuidedReviewDesigner
through the real dispatcher at /GuidedReviewDemo with zero page errors.
